requirements.txt
fastapi
uvicorn
openai-clip
torch
torchvision
pillow
cloudinary
python-multipart
requests
python-dotenv
numpy
